Allow decimal shop prices

+++ app/(app)/dash/makers/page.tsx
@@ -39,11 +39,11 @@ export default async function MakersPage({
       approvedMinutes: sql<number>`coalesce((
         select sum(${projects.approvedMinutes}) from ${projects}
         where ${projects.userSub} = ${users.sub} and ${projects.decision} = 'approved'
-      ), 0)::int`,
+      ), 0)`,
       beans: sql<number>`coalesce((
         select sum(${beansLedger.delta}) from ${beansLedger}
         where ${beansLedger.userSub} = ${users.sub}
-      ), 0)::int`,
+      ), 0)`,
       orderCount: sql<number>`(
         select count(*) from ${orders} where ${orders.userSub} = ${users.sub}
       )::int`,
